Cambio Final

+++ ExamenTercerParcial/src/main/java/ec/edu/espe/arqsoftware/examen/zuniga/dao/PagoPrestamoRepository.java
@@ -3,5 +3,5 @@ package ec.edu.espe.arqsoftware.examen.zuniga.dao;
 import ec.edu.espe.arqsoftware.examen.zuniga.model.PagoPrestamo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagoPrestamoRepository extends JpaRepository<PagoPrestamo, Long> {
+public interface PagoPrestamoRepository extends JpaRepository<PagoPrestamo, Integer> {
 }
